+++ emulate_android.sh
@@ -46,7 +46,7 @@ echo 'export PATH=$PATH:$ANDROID_HOME/emulator' >> ~/.profile
 avdmanager create avd -n "android-small" -k "system-images;android-28;google_apis_playstore;x86_64"
 avdmanager create avd -n Emulator-Api23-Default -c 12M -f -k "system-images;android-28;google_apis_playstore;x86_64"
 
-# Show installed android devices:
+# Show installed android devs:
 emulator -list-avds
 
 # Launch emulated phone:
