+++ src/apkcontroller/org_torproject_android/V16_6_3_RC_1/Apk_script.py
@@ -3,11 +3,10 @@
 Android names this app: org.torproject.android
 """
 
-from typing import TYPE_CHECKING, Callable, Dict, List, Optional
+from typing import TYPE_CHECKING, Dict, List, Optional
 
 import networkx as nx
 from typeguard import typechecked
-from uiautomator import AutomatorDevice
 
 from src.apkcontroller.create_screens import create_screens
 from src.apkcontroller.org_torproject_android.V16_6_3_RC_1.screen_flow import (
@@ -66,22 +65,3 @@ class Apk_script:
                 script_graph.nodes[nodename]["is_start"] = True
             else:
                 script_graph.nodes[nodename]["is_start"] = False
-
-    @typechecked
-    def perform_action(
-        self,
-        device: AutomatorDevice,
-        next_actions: Callable,
-        additional_info: Dict,
-    ) -> Dict:
-        """Performs the first action list in the list of action lists."""
-        action_output: Dict = next_actions(
-            device=device,
-            additional_info=additional_info,
-        )
-        if "expected_screens" not in action_output.keys():
-            raise KeyError(
-                "Error, the action output did not contain the expected "
-                "screens."
-            )
-        return action_output
