+++ src/apkcontroller/org_torproject_android/V16_6_3_RC_1/screen_4.py
@@ -67,12 +67,12 @@ def screen_4(script_description: Dict) -> Screen:
 
 # pylint: disable=W0613
 @typechecked
-def actions_0(device: AutomatorDevice, additional_info: Dict) -> Dict:
+def actions_0(dev: AutomatorDevice, additional_info: Dict) -> Dict:
     """Performs the actions in option 0 in this screen.
 
     For this screen, it clicks the "DONE" button in the bottom right.
     """
-    device(resourceId="org.torproject.android:id/done").click()
+    dev(resourceId="org.torproject.android:id/done").click()
 
     # Return the expected screens, using get_expected_screen_nrs.
     action_nr: int = int(inspect.stack()[0][3][8:])
