+++ src/apkcontroller/run_script.py
@@ -1,28 +1,28 @@
 """Starts a script to control an app."""
 
 
-from typing import Dict, List
+from typing import Callable, Dict, List
 
 from typeguard import typechecked
-from uiautomator import device
+from uiautomator import AutomatorDevice
 
 from src.apkcontroller.helper import export_screen_data_if_valid, launch_app
+from src.apkcontroller.org_torproject_android.plot_script_flow import (
+    visualise_script_flow,
+)
 from src.apkcontroller.org_torproject_android.V16_6_3_RC_1.Apk_script import (
     Apk_script,
-)
-from src.apkcontroller.org_torproject_android.V16_6_3_RC_1.screen_flow import (
-    visualise_script_flow,
 )
 from src.apkcontroller.script_helper import can_proceed, get_start_nodes
 
 
 @typechecked
-def run_script(script: Apk_script) -> None:
+def run_script(script: Apk_script, dev: AutomatorDevice) -> None:
     """Runs the incoming script on the phone.
 
     Script folder structure: src/apkcontroller/app_name/version.py with
     app_name is something like: com_whatsapp_android (not: Whatsapp). It
-    is derived from how your android device calls the app, with the dots
+    is derived from how your android dev calls the app, with the dots
     replaced by underscores. E.g. com.whatsapp.android or something like
     that.
     """
@@ -41,7 +41,7 @@ def run_script(script: Apk_script) -> None:
     expected_screens: List[int] = get_start_nodes(script.script_graph)
 
     _, screen_nr = can_proceed(
-        device=device,
+        dev=dev,
         expected_screennames=expected_screens,
         retry=True,
         script=script,
@@ -53,7 +53,7 @@ def run_script(script: Apk_script) -> None:
     # known already.
     while next_actions is not None:
         _, screen_nr = can_proceed(
-            device=device,
+            dev=dev,
             expected_screennames=expected_screens,
             retry=retry,
             script=script,
@@ -63,9 +63,9 @@ def run_script(script: Apk_script) -> None:
         print(f"screen_nr={screen_nr}")
 
         # Export the data of the screens if they happen to be found in the
-        # device already.
+        # dev already.
         export_screen_data_if_valid(
-            device=device,
+            dev=dev,
             overwrite=script.script_description["overwrite"],
             screens=[screen],
         )
@@ -86,8 +86,8 @@ def run_script(script: Apk_script) -> None:
             additional_info["script_graph"] = script.script_graph
 
             # Perform the actual action.
-            action_output: Dict = script.perform_action(
-                device=device,
+            action_output: Dict = perform_action(
+                dev=dev,
                 next_actions=next_actions,
                 additional_info=additional_info,
             )
@@ -95,3 +95,21 @@ def run_script(script: Apk_script) -> None:
             script.script_description["past_screens"].append(screen_nr)
 
     print(f'Done with script:{script.script_description["app_name"]}')
+
+
+@typechecked
+def perform_action(
+    dev: AutomatorDevice,
+    next_actions: Callable,
+    additional_info: Dict,
+) -> Dict:
+    """Performs the first action list in the list of action lists."""
+    action_output: Dict = next_actions(
+        dev=dev,
+        additional_info=additional_info,
+    )
+    if "expected_screens" not in action_output.keys():
+        raise KeyError(
+            "Error, the action output did not contain the expected screens."
+        )
+    return action_output
